subrace plugin update #1
This commit starts to fill-out shayan's subrace engine to 3.0.6.9.  It
is used as an example of how plugins work in the core framework,
allowing various systems to be sequestered/containerized so that still
work across the module, but live in their own folders, making it easy to
replace, update, find or remove the entire system at once.  The subrace
plugin contains all items, placeables, scripts, and other resources
required to run the system.